services/voice-recognition-service/requirements.txt
grpcio==1.71.0
grpcio-reflection==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
numpy==2.0.2
sentence-transformers==4.1.0
transformers==4.52.3
torch==2.7.0
torchaudio==2.7.0
soundfile==0.13.1
sentencepiece==0.2.0
faster-whisper==1.1.1
